Add PUT endpoint TODO:  -create get single user endpoint for admin  -refactor of the code written  -convert utils into factory
