+++ src/main/java/com/github/WeeiaEduTeam/InfinityFinanceAPI/appuser/dto/CreateAppUserAdminDTO.java
@@ -1,9 +1,7 @@
 package com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.dto;
 
-import com.github.WeeiaEduTeam.InfinityFinanceAPI.role.Role;
-import com.github.WeeiaEduTeam.InfinityFinanceAPI.role.role.RoleDTO;
+import com.github.WeeiaEduTeam.InfinityFinanceAPI.role.dto.RoleDTO;
 import lombok.*;
-import lombok.experimental.FieldNameConstants;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.List;
