+++ src/main/java/com/github/WeeiaEduTeam/InfinityFinanceAPI/appuser/dto/ReplaceAppUserAllDetailsDTO.java
@@ -1,6 +1,6 @@
 package com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.dto;
 
-import com.github.WeeiaEduTeam.InfinityFinanceAPI.role.role.RoleDTO;
+import com.github.WeeiaEduTeam.InfinityFinanceAPI.role.dto.RoleDTO;
 import lombok.*;
 import lombok.experimental.FieldNameConstants;
 
@@ -8,17 +8,11 @@ import java.util.List;
 
 @Getter
 @Setter
-@Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @FieldNameConstants
-public class ReplaceAppUserAllDetailsDTO {
+public class ReplaceAppUserAllDetailsDTO  extends CreateAppUserAdminDTO {
 
-    private String username;
-    private String email;
-    private String password;
     private String firstName;
     private String secondName;
-
-    private List<RoleDTO> roles;
 }
