+++ src/main/java/com/github/WeeiaEduTeam/InfinityFinanceAPI/appuser/dto/ReplaceAppUserByUserDTO.java
@@ -1,10 +1,7 @@
 package com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.dto;
 
-import com.github.WeeiaEduTeam.InfinityFinanceAPI.role.role.RoleDTO;
 import lombok.*;
 import lombok.experimental.FieldNameConstants;
-
-import java.util.List;
 
 @Getter
 @Setter
