+++ src/main/java/com/github/WeeiaEduTeam/InfinityFinanceAPI/appuser/strategy/AppUserAddAdminRole.java
@@ -1,30 +1,28 @@
-package com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.strategies;
+package com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.strategy;
 
 import com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.AppUser;
 import com.github.WeeiaEduTeam.InfinityFinanceAPI.role.Role;
 import com.github.WeeiaEduTeam.InfinityFinanceAPI.role.RoleService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.github.WeeiaEduTeam.InfinityFinanceAPI.role.RoleUtil.ROLE_USER;
+import static com.github.WeeiaEduTeam.InfinityFinanceAPI.role.RoleUtil.ROLE_ADMIN;
 
 @Component
-public class AppUserAddUserRole implements AppUserRoleStrategy {
+public class AppUserAddAdminRole implements AppUserRoleStrategy {
 
-    private static AppUserAddUserRole instance = null;
+    private static AppUserAddAdminRole instance = null;
     private final RoleService roleService;
 
-    private AppUserAddUserRole(RoleService roleService) {
+    private AppUserAddAdminRole(RoleService roleService) {
         this.roleService = roleService;
     }
 
-    public static AppUserAddUserRole getInstance(RoleService roleService) {
+    public static AppUserAddAdminRole getInstance(RoleService roleService) {
         if (instance == null) {
-            instance = new AppUserAddUserRole(roleService);
+            instance = new AppUserAddAdminRole(roleService);
         }
 
         return instance;
@@ -34,8 +32,8 @@ public class AppUserAddUserRole implements AppUserRoleStrategy {
     public void addRolesForUser(AppUser user) {
         var roles = getRoles(user);
 
-        if(!userRoleExists(roles)) {
-            var userRoles = getUserRoleList();
+        if(!adminRoleExists(roles)) {
+            var userRoles = getAdminRoleList();
 
             user.setRoles(userRoles);
         }
@@ -48,15 +46,16 @@ public class AppUserAddUserRole implements AppUserRoleStrategy {
         return user.getRoles();
     }
 
-    private boolean userRoleExists(List<Role> roles) {
+    private boolean adminRoleExists(List<Role> roles) {
         return roles.stream()
                 .map(Role::getName)
-                .filter(e -> e.equals(ROLE_USER))
+                .filter(e -> e.equals(ROLE_ADMIN))
                 .toList()
                 .size() > 0;
     }
 
-    private List<Role> getUserRoleList() {
-        return List.of(roleService.getUserRoleOrCreate());
+
+    private List<Role> getAdminRoleList() {
+        return List.of(roleService.getUserRoleOrCreate(), roleService.getAdminRoleOrCreate());
     }
 }
