+++ src/main/java/com/github/WeeiaEduTeam/InfinityFinanceAPI/appuser/strategy/AppUserRoleStrategy.java
@@ -1,7 +1,6 @@
-package com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.strategies;
+package com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.strategy;
 
 import com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.AppUser;
-import com.github.WeeiaEduTeam.InfinityFinanceAPI.role.RoleService;
 
 public interface AppUserRoleStrategy {
     void addRolesForUser(AppUser user);
