+++ src/main/java/com/github/WeeiaEduTeam/InfinityFinanceAPI/role/RoleService.java
@@ -1,10 +1,13 @@
 package com.github.WeeiaEduTeam.InfinityFinanceAPI.role;
 
 import com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.AppUser;
-import com.github.WeeiaEduTeam.InfinityFinanceAPI.role.role.RoleDTO;
+import com.github.WeeiaEduTeam.InfinityFinanceAPI.appuser.AppUserRepository;
+import com.github.WeeiaEduTeam.InfinityFinanceAPI.role.dto.RoleDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
 
 import static com.github.WeeiaEduTeam.InfinityFinanceAPI.role.RoleUtil.ROLE_ADMIN;
 import static com.github.WeeiaEduTeam.InfinityFinanceAPI.role.RoleUtil.ROLE_USER;
@@ -53,11 +56,11 @@ public class RoleService {
         return roleUtil.mapRoleToRoleDTO(role);
     }
 
-    public void deleteRoleFromUser(AppUser user) {
-        var roles = user.getRoles();
+    public void deleteRolesFromUser(AppUser user) {
+        int roleLength = user.getRoles().size();
 
-        roles.stream()
-                .map(e -> getRole(e.getName()))
-                .forEach(e -> e.getUsers().remove(user));
+        if (roleLength > 0) {
+            user.getRoles().subList(0, roleLength).clear();
+        }
     }
 }
